+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+- Add the ability to set an `alwaysOn` property for filters. This should be a boolean and will hide the close button and automatically show the filter.
+- Add the ability to remove the filter from the dropdown and hide the close button using `once: true`.
+- Add search functionality to the list dsl. This will search on the `title` field or specified fields if set.
+
 ## v1.0.0-9.1.0 (9 August 2017)
 
 ### BREAKING CHANGES